refactor: diary 패키지에 있던 내용들 todo 패키지에 통합
