+++ src/main/java/com/hanghae/bulletbox/category/dto/ResponseCategoryDto.java
@@ -1,6 +1,6 @@
 package com.hanghae.bulletbox.category.dto;
 
-import com.hanghae.bulletbox.diary.dto.DailyDto;
+import com.hanghae.bulletbox.todo.dto.DailyDto;
 
 import lombok.AccessLevel;
 import lombok.Getter;
