+++ src/main/java/com/hanghae/bulletbox/todo/dto/RequestLoadFavoriteDto.java
@@ -1,4 +1,4 @@
-package com.hanghae.bulletbox.diary.dto;
+package com.hanghae.bulletbox.todo.dto;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
