+++ src/main/java/com/hanghae/bulletbox/todo/dto/TodoDto.java
@@ -2,7 +2,6 @@ package com.hanghae.bulletbox.todo.dto;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 
-import com.hanghae.bulletbox.diary.dto.DailyTodoDto;
 import com.hanghae.bulletbox.favorite.dto.FavoriteDto;
 import com.hanghae.bulletbox.member.dto.MemberDto;
 import com.hanghae.bulletbox.member.entity.Member;
